backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.29.0
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
boto3==1.34.0
django-storages==1.14.2
